+++ WalletWasabi.Fluent/Controls/DestinationEntry/Components/QrCodeButton.axaml.cs
